Daily Algo Leetcode Medium
Algo 120 Triangle.
Given a nested triagular array, return the min sum path, while only
hoping adjcently from row to row.
I was able to solve this algo in three different ways. Initally I wanted
to solve this in a top down method starting at triangle[0][0], then walk
down in the most logical min path.

But I did not take into account that the immedate path may not be the
most efficient, therefore I must exhaust all the options. Upon
submission I found that the top-down solution was not the most optimal
for all permutations.

Then by appraching it from a different angle and moving from penultimate
row-up. I was able to add "from below" still choosing the min number to
float/add to the top. Therefore at the end of the algo I only had to
return the tip of the triangle to complete the algo.