Daily Algo Leetcode BST
Completed Leetcode's algo 543 Diameter of Binary Tree.

Double recrusive function. I have to helper functions, one that
calculates the height of the given TreeNode, and the other helper
function that add the left and right heights of the nodes subtrees to
find the diameter/distance between any two nodes. Then may override the
golbal variable max to be returened at the end of the function.